+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## 3.4.7 - 2021-11-22
+- Added `--timeout=<time in seconds>` to the `wp s1 queues process` command
+- Updated Queue Tasks to be created by the container so that dependency injection works.
+- Update Queues documentation with examples.
+
 ## 3.4.6 - 2021-09-10
 
 ## 3.4.5 - 2021-09-10